+++ lib/jxl/enc_xyb.cc
@@ -1,16 +1,7 @@
-// Copyright (c) the JPEG XL Project
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 #include "lib/jxl/enc_xyb.h"
 
@@ -88,7 +79,7 @@ void TestCubeRoot() {
   float max_err = 0.0f;
   for (uint64_t x5 = 0; x5 < 2000000; x5++) {
     const float x = x5 * 1E-5f;
-    const float expected = std::cbrt(x);
+    const float expected = cbrtf(x);
     HWY_ALIGN float approx[MaxLanes(d)];
     Store(CubeRootAndAdd(Set(d, x), Zero(d)), d, approx);
 
@@ -282,7 +273,7 @@ const ImageBundle* ToXYB(const ImageBundle& in, ThreadPool* pool,
     Store(absorb, d, premul_absorb + i * N);
   }
   for (size_t i = 0; i < 3; ++i) {
-    const auto neg_bias_cbrt = Set(d, -std::cbrt(kOpsinAbsorbanceBias[i]));
+    const auto neg_bias_cbrt = Set(d, -cbrtf(kOpsinAbsorbanceBias[i]));
     Store(neg_bias_cbrt, d, premul_absorb + (9 + i) * N);
   }
