+++ lib/jxl/linalg.cc
@@ -1,16 +1,7 @@
-// Copyright (c) the JPEG XL Project
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 #include "lib/jxl/linalg.h"
 
@@ -56,7 +47,7 @@ void GivensRotation(const double x, const double y, double* c, double* s) {
     *c = x < 0.0 ? -1.0 : 1.0;
     *s = 0.0;
   } else {
-    const double h = std::hypot(x, y);
+    const double h = hypot(x, y);
     const double d = 1.0 / h;
     *c = x * d;
     *s = -y * d;
@@ -127,7 +118,7 @@ double WilkinsonShift(const double a0, const double a1, const double b) {
     return a1 - std::abs(b);
   }
   const double sign_d = d > 0.0 ? 1.0 : -1.0;
-  return a1 - b * b / (d + sign_d * std::hypot(d, b));
+  return a1 - b * b / (d + sign_d * hypotf(d, b));
 }
 
 void ImplicitQRStep(ImageD* const JXL_RESTRICT U, double* const JXL_RESTRICT a,
