+++ lib/jxl/quant_weights_test.cc
@@ -1,16 +1,7 @@
-// Copyright (c) the JPEG XL Project
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 #include "lib/jxl/quant_weights.h"
 
 #include <stdlib.h>
@@ -23,7 +14,7 @@
 #include <random>
 
 #include "lib/jxl/dct_for_test.h"
-#include "lib/jxl/dec_transforms.h"
+#include "lib/jxl/dec_transforms_testonly.h"
 #include "lib/jxl/enc_modular.h"
 #include "lib/jxl/enc_quant_weights.h"
 #include "lib/jxl/enc_transforms.h"
@@ -203,9 +194,8 @@ TEST_P(QuantWeightsTargetTest, DCTUniform) {
 
     for (size_t i = 0; i < 64; i++) {
       // DCTSlow doesn't multiply/divide by 1/N, so we do it manually.
-      slow_coeffs[i] =
-          std::round(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
-      coeffs[i] = std::round(coeffs[i] / dequant_matrices.Matrix(dct, 0)[i]) *
+      slow_coeffs[i] = roundf(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
+      coeffs[i] = roundf(coeffs[i] / dequant_matrices.Matrix(dct, 0)[i]) *
                   dequant_matrices.Matrix(dct, 0)[i];
     }
     IDCTSlow<8>(slow_coeffs);
@@ -227,9 +217,8 @@ TEST_P(QuantWeightsTargetTest, DCTUniform) {
     DCTSlow<16>(slow_coeffs);
 
     for (size_t i = 0; i < 64 * 4; i++) {
-      slow_coeffs[i] =
-          std::round(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
-      coeffs[i] = std::round(coeffs[i] / dequant_matrices.Matrix(dct, 0)[i]) *
+      slow_coeffs[i] = roundf(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
+      coeffs[i] = roundf(coeffs[i] / dequant_matrices.Matrix(dct, 0)[i]) *
                   dequant_matrices.Matrix(dct, 0)[i];
     }
